backend/dashboard/serializers.py
from rest_framework import serializers
from .models import Product, Post

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'title', 'description', 'features', 'price', 'image']


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['id', 'title', 'content', 'created_at', 'image']  # Menambahkan field image
